+++ test/lib/compile/openapi.test.js
@@ -1,3 +1,4 @@
+const { path } = require('@sap/cds/lib/utils/cds-utils');
 const toOpenApi = require('../../../lib/compile');
 const cds = require('@sap/cds')
 
@@ -235,6 +236,34 @@ describe('OpenAPI export', () => {
     catch (e) {
       expect(e.message).toBe("The input server object is invalid.");
     }
+  });
+
+  test('options: config-file - without inline options', () => {
+    const csn = cds.compile.to.csn(`
+      service A {entity E { key ID : UUID; };};`
+    );
+    const openapi = toOpenApi(csn, { 'openapi:config-file': path.resolve("./test/lib/compile/data/configFile.json") });
+    expect(openapi.servers).toBeTruthy();
+    expect(openapi).toMatchObject({ servers: [{ url: 'http://foo.bar:8080/rest/A' }, { url: "http://foo.bar:8080/a/foo/rest/A" }] });
+    expect(openapi.info.description).toMatch(/yuml.*diagram/i); 
+    expect(openapi['x-odata-version']).toMatch('4.1');
+  });
+
+  test('options: config-file - with inline options, inline options given precedence', () => {
+    const csn = cds.compile.to.csn(`
+       @title:'It is located at http://example.com:8080' service A {entity E { key ID : UUID;};};`
+    );
+    const options = {
+      'openapi:config-file': path.resolve("./test/lib/compile/data/configFile.json"),
+      'openapi:url': "http://example.com:8080",
+      'odata-version': '4.0',
+      'openapi:diagram': "false"
+    }
+    const openapi = toOpenApi(csn, options);
+    expect(openapi.info.title).toMatch(/http:\/\/example.com:8080/i)
+    expect(openapi.info.description).not.toMatch(/yuml.*diagram/i);
+    expect(openapi['x-odata-version']).toMatch('4.0');
+    expect(openapi).toMatchObject({ servers: [{ url: 'http://foo.bar:8080/odata/v4/A' }, { url: "http://foo.bar:8080/a/foo/odata/v4/A" }] });
   });
 
   test('annotations: root entity property', () => {